Add files via upload
This crisis simulating VR-experience includes two scenes, one depicts a room with a table that displays items the user can pick to help them in the crisis, the other depicts a snowstorm where the user shall navigate to the nearest shelter. Ideally, the outcome of the experience should differ depending on which items the user picks. For example, a parachute will not be as useful as matches to keep you warm in a snowstorm. By allowing the user to re-experience the simulation and see first hand what items may be more (or less) useful in a crisis situation, our aim is that the user will have a better idea of what to bring if/when an actual crisis emerges. However, this degree of interactivity has
not yet been implemented, meaning the experience will be the same regardless of which items are selected, at this time.

The experience will be able to be adjusted with various settings, but the purpose of it all is to make it feel safe and secure to experience this crisis. Therefore, we have chosen not to put time pressure on the user but instead encourage them to perform the tasks in a proper and safe manner.